+++ lib/blocs/password/password_bloc.dart
@@ -28,7 +28,7 @@ class PasswordBloc extends Bloc<PasswordEvent, PasswordState> {
     try {
       final cards = await _vaultService.getAllCards();
       emit(PasswordLoaded(cards: cards));
-    } catch (e) {
+    } on Object catch (e) {
       emit(PasswordError(e.toString()));
     }
   }
@@ -43,7 +43,7 @@ class PasswordBloc extends Bloc<PasswordEvent, PasswordState> {
     try {
       final cards = await _vaultService.search(event.query);
       emit(PasswordLoaded(cards: cards, query: event.query));
-    } catch (e) {
+    } on Object catch (e) {
       emit(PasswordError(e.toString()));
     }
   }
@@ -56,7 +56,7 @@ class PasswordBloc extends Bloc<PasswordEvent, PasswordState> {
     try {
       await _vaultService.createCard(event.payload);
       add(PasswordLoadRequested()); // Reload list
-    } catch (e) {
+    } on Object catch (e) {
       emit(PasswordError(e.toString()));
     }
   }
@@ -69,7 +69,7 @@ class PasswordBloc extends Bloc<PasswordEvent, PasswordState> {
     try {
       await _vaultService.updateCard(event.cardId, event.payload);
       add(PasswordLoadRequested()); // Reload list
-    } catch (e) {
+    } on Object catch (e) {
       emit(PasswordError(e.toString()));
     }
   }
@@ -82,7 +82,7 @@ class PasswordBloc extends Bloc<PasswordEvent, PasswordState> {
     try {
       await _vaultService.deleteCard(event.cardId);
       add(PasswordLoadRequested()); // Reload list
-    } catch (e) {
+    } on Object catch (e) {
       emit(PasswordError(e.toString()));
     }
   }