+++ test/helpers/test_helpers.dart
@@ -2,7 +2,7 @@
 ///
 /// 在所有测试中统一 import 此文件即可获得所有工具：
 ///   import 'helpers/test_helpers.dart';
-library test_helpers;
+library;
 
 export 'test_fixtures.dart';
 export 'test_matchers.dart';